bridge/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
minio
redis
requests
numpy
pillow
python-multipart
